Id property

+++ Reciplease/Reciplease/Model/FavoriteRecipesRepository.swift
@@ -8,49 +8,80 @@
 
 import Foundation
 import CoreData
+import UIKit
 
 protocol FavoriteRecipesRepository {
-    func getFavoriteRecipes() -> [FavoriteRecipe]
-    func addRecipeToFavoriteFromDetails(recipe: Recipe)
-    func addRecipeToFavoriteFromFavorite(recipe: FavoriteRecipe)
-    func removeRecipeOfFavorites(recipe: FavoriteRecipe)
+    func getFavoriteRecipes() throws -> [FavoriteRecipe]
+    func getFavoriteRecipe(by uri: String) throws-> FavoriteRecipe?
+    func addRecipeToFavorite(totalTime: Int, image: String, label: String, ingredientLines: [String], uri: String)
+    func removeRecipe(by uri: String) throws
+}
+
+enum FavoriteRecipeRequestError: Error {
+    case requestForFavoriteRecipesError
+    case requestForGettingRecipeByUriError
+    case removingRecipeError
+    
 }
 
 class FavoriteRecipesRepositoryImplementation: FavoriteRecipesRepository {
     
-    func getFavoriteRecipes() -> [FavoriteRecipe] {
+    func getFavoriteRecipes() throws -> [FavoriteRecipe] {
         var favorites: [FavoriteRecipe] = []
         let fetchRequest: NSFetchRequest<FavoriteRecipe> = FavoriteRecipe.fetchRequest()
         do {
             let favoritesRecipes = try PersistenceService.context.fetch(fetchRequest)
             favorites = favoritesRecipes
         } catch {
-            
+            throw FavoriteRecipeRequestError.requestForFavoriteRecipesError
         }
+        
         return favorites
     }
     
-    func addRecipeToFavoriteFromDetails(recipe: Recipe) {
-        let favoriteRecipe = FavoriteRecipe(context: PersistenceService.context)
-        favoriteRecipe.totalTime = recipe.totalTime
-        favoriteRecipe.image = recipe.image
-        favoriteRecipe.label = recipe.label
-        favoriteRecipe.ingredientLines = recipe.ingredientLines
-        PersistenceService.saveContext()
+    func getFavoriteRecipe(by uri: String) throws -> FavoriteRecipe? {
+        var favoriteRecipe: FavoriteRecipe?
+        let fetchRequest: NSFetchRequest<FavoriteRecipe> = FavoriteRecipe.fetchRequest()
+        do {
+            let favoritesRecipes = try PersistenceService.context.fetch(fetchRequest)
+            favoriteRecipe = favoritesRecipes.first { favoriteRecipe -> Bool in
+                favoriteRecipe.uri == uri
+            }
+        } catch {
+            throw FavoriteRecipeRequestError.requestForGettingRecipeByUriError
+        }
+        
+        return favoriteRecipe
+        
     }
     
-    func addRecipeToFavoriteFromFavorite(recipe: FavoriteRecipe) {
+    func addRecipeToFavorite(totalTime: Int, image: String, label: String, ingredientLines: [String], uri: String) {
         let favoriteRecipe = FavoriteRecipe(context: PersistenceService.context)
-        favoriteRecipe.totalTime = recipe.totalTime
-        favoriteRecipe.image = recipe.image
-        favoriteRecipe.label = recipe.label
-        favoriteRecipe.ingredientLines = recipe.ingredientLines
+        favoriteRecipe.totalTime = totalTime
+        favoriteRecipe.image = image
+        favoriteRecipe.label = label
+        favoriteRecipe.ingredientLines = ingredientLines
+        favoriteRecipe.uri = uri
         PersistenceService.saveContext()
     }
+
     
     
-    func removeRecipeOfFavorites(recipe: FavoriteRecipe) {
-        PersistenceService.context.delete(recipe)
-        PersistenceService.saveContext()
+    func removeRecipe(by uri: String) throws {
+        var favorites: [FavoriteRecipe] = []
+        let fetchRequest: NSFetchRequest<FavoriteRecipe> = FavoriteRecipe.fetchRequest()
+        do {
+            let favoritesRecipes = try PersistenceService.context.fetch(fetchRequest)
+            favorites = favoritesRecipes
+            for recipe in favorites {
+                if recipe.uri == uri {
+                    PersistenceService.context.delete(recipe)
+                    PersistenceService.saveContext()
+                }
+            }
+        } catch {
+            throw FavoriteRecipeRequestError.removingRecipeError
+        }
+        
     }
 }
